Add Block Breaking particle
Monty moles now emit a block break particle upon leaving
the ground and play a sound. Also AI has improved to be
very similar to the original game's.

Also added a dust particle when the mole switches directions,
exactly the same as the player does.

Also fixed debug line drawing and disabled it.