Add delay to koopa head stomp sound
Koopas play two sounds when you stomp on their
heads (with a spin jump), first a sound right away,
then a sound after you have finished a chain of them.

This is accomplished using a timer which is reset
everytime a new koopa is stomped on. When it reaches 0,
the second sound effect is played.