Fix Game session bug
Prior to this commit, game sessions would
occasionally be recorded incorrectly, this has
now been fixed.

The problem was that the same member variable
was being used to show sessions on the screen and
record the current play through.

There are now two members, one for showing,
one for recording.