Add Message class
Messages are now stored as strings, rather
than images.

Message blocks used to handle the ticking/painting
of messages to the screen, but yoshi can also put
a message on the screen so this functionality has
now been moved to the Message class.

SMWFont can now correctly handle more characters.